Add support of always() function

+++ neuro_flow/batch_executor.py
@@ -21,7 +21,7 @@ from .commands import CmdProcessor
 from .config_loader import BatchRemoteCL
 from .context import BatchActionContext, BatchContext, DepCtx, NeedsCtx, TaskContext
 from .storage import Attempt, BatchStorage, FinishedTask, SkippedTask, StartedTask
-from .types import FullID, LocalPath, TaskStatus
+from .types import AlwaysT, FullID, LocalPath, TaskStatus
 from .utils import TERMINATED_JOB_STATUSES, fmt_id, fmt_raw_id, fmt_status
 
 
@@ -73,6 +73,7 @@ class BatchExecutor:
         self._started: Dict[FullID, StartedTask] = {}
         self._finished: Dict[FullID, FinishedTask] = {}
         self._skipped: Dict[FullID, SkippedTask] = {}
+        self._is_cancelling = False
 
         # A note about default value:
         # AS: I have no idea what timeout is better;
@@ -107,9 +108,12 @@ class BatchExecutor:
 
             click.echo("Find last attempt")
             attempt = await self._storage.find_attempt(bake)
-            if attempt.result in TERMINATED_JOB_STATUSES:
+            if attempt.result == JobStatus.CANCELLED:
+                self._is_cancelling = True
+            elif attempt.result in TERMINATED_JOB_STATUSES:
                 str_attempt_status = fmt_status(attempt.result)
                 click.echo(f"Attempt #{attempt.number} is already {str_attempt_status}")
+                return
 
             click.echo(f"Fetch attempt #{attempt.number}")
             (
@@ -138,45 +142,38 @@ class BatchExecutor:
                             self._topos[new_prefix] = (new_ctx, new_topo, [])
 
                 ok = await self._process_started(attempt)
-                if not ok:
-                    await self._do_cancellation(attempt)
-                    break
 
                 # Check for cancellation
-                attempt = await self._storage.find_attempt(bake, attempt.number)
-                if attempt.result == JobStatus.CANCELLED:
-                    await self._do_cancellation(attempt)
-                    click.echo(
-                        f"Attempt #{attempt.number} {fmt_status(JobStatus.CANCELLED)}"
-                    )
-                    return
+                if not self._is_cancelling:
+                    attempt = await self._storage.find_attempt(bake, attempt.number)
+
+                    if not ok or attempt.result == JobStatus.CANCELLED:
+                        self._is_cancelling = True
+                        await self._stop_running()
 
                 await asyncio.sleep(self._polling_timeout)
 
             attempt_status = self._accumulate_result()
             str_attempt_status = fmt_status(attempt_status)
-            await self._storage.finish_attempt(attempt, attempt_status)
+            if attempt.result not in TERMINATED_JOB_STATUSES:
+                await self._storage.finish_attempt(attempt, attempt_status)
             click.echo(f"Attempt #{attempt.number} {str_attempt_status}")
 
-    async def _do_cancellation(self, attempt: Attempt) -> None:
-        killed = []
+    async def _always_enabled(self, full_id: FullID) -> bool:
+        prefix, tid = full_id[:-1], full_id[-1]
+        ctx, _, _ = self._topos[prefix]
+        needs = self._build_needs(prefix, ctx.graph[full_id])
+        return await ctx.is_enabled(tid, needs=needs) is AlwaysT()
+
+    async def _stop_running(self) -> None:
         for st in self._started.values():
-            if st.raw_id and st.id not in self._finished:
+            if (
+                st.raw_id
+                and st.id not in self._finished
+                and not await self._always_enabled(st.id)
+            ):
+                click.echo(f"Task {fmt_id(st.id)} is being killed")
                 await self._client.jobs.kill(st.raw_id)
-                killed.append(st.id)
-        while any(k_id not in self._finished for k_id in killed):
-            await self._process_started(attempt)
-            await asyncio.sleep(self._polling_timeout)
-        # All jobs stopped, mark as canceled started actions
-        for st in self._started.values():
-            if st.id not in self._finished:
-                assert not st.raw_id
-                await self._storage.finish_batch_action(
-                    attempt,
-                    self._next_task_no(),
-                    st,
-                    DepCtx(TaskStatus.CANCELLED, {}),
-                )
 
     def _next_task_no(self) -> int:
         return len(self._started) + len(self._finished) + len(self._skipped)
@@ -210,7 +207,8 @@ class BatchExecutor:
             tid = full_id[-1]
             needs = self._build_needs(prefix, ctx.graph[full_id])
             assert full_id[:-1] == prefix
-            if not await ctx.is_enabled(tid, needs=needs):
+            enabled = await ctx.is_enabled(tid, needs=needs)
+            if not enabled or (self._is_cancelling and enabled is not AlwaysT()):
                 # Make task started and immediatelly skipped
                 skipped_task = await self._skip_task(
                     attempt, self._next_task_no(), full_id
@@ -287,7 +285,9 @@ class BatchExecutor:
                 assert isinstance(ctx, BatchActionContext)
 
                 needs = self._build_needs(ctx.prefix, await ctx.get_output_needs())
-                outputs = await ctx.calc_outputs(needs)
+                outputs = await ctx.calc_outputs(
+                    needs, is_cancelling=self._is_cancelling
+                )
 
                 ft = await self._storage.finish_batch_action(
                     attempt,
@@ -317,7 +317,7 @@ class BatchExecutor:
         for full_id in deps:
             dep_id = full_id[-1]
             if full_id in self._skipped:
-                needs[dep_id] = DepCtx(TaskStatus.DISABLED, {})
+                needs[dep_id] = DepCtx(TaskStatus.SKIPPED, {})
             else:
                 dep = self._finished.get(full_id)
                 if dep is None:
